% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_regressions.R
\name{analyse_regressions}
\alias{analyse_regressions}
\title{Perform Linear Regression Analysis Between Density and Chemical Layers}
\usage{
analyse_regressions(density_index, compare_indices)
}
\arguments{
\item{density_index}{Integer. The index of the layer in `result_array` that represents the density.}

\item{compare_indices}{Integer vector. A vector of indices corresponding to the layers in `result_array` to compare against the density layer.}
}
\value{
A data frame with two columns: \code{Couche}, indicating the index of the compared layer, and \code{R_squared}, representing the coefficient of determination for each regression analysis.
}
\description{
This function performs linear regression analysis between a specified density layer and multiple chemical layers within a 3-dimensional array, generating interactive regression plots using Plotly.
}
\details{
The function extracts the specified density layer and each chemical layer from `result_array` using the provided indices. It filters the data to include only positive values, performs linear regression analysis, and plots the regression results interactively using Plotly. The function outputs a data frame containing the R-squared values for each regression analysis.
}
\examples{
\dontrun{
# Assuming `result_array` is a 3D array with suitable data
regression_results <- analyse_regressions(density_index = 5, compare_indices = c(6, 7, 8))
print(regression_results)
}

}
