% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_matrix.R
\name{classify_matrix}
\alias{classify_matrix}
\title{Classify a Matrix Layer Based on Noniles}
\usage{
classify_matrix(layer_index)
}
\arguments{
\item{layer_index}{Integer. The index of the layer to be classified. Must be between 1 and the number of layers in the array.}
}
\value{
A matrix of the same dimensions as the input layer, with numeric values representing the classified categories.
}
\description{
This function classifies values in a specified layer of a 3-dimensional array into categories based on noniles.
}
\details{
The function extracts a specified layer from a 3-dimensional array named `result_array` and classifies each value based on noniles. It assigns labels such as "Very Low", "Medium", "High", etc., to classify the values into ten distinct categories.
}
\examples{
\dontrun{
# Assuming `result_array` is a 3D array with suitable dimensions
classified_matrix <- classify_matrix(layer_index = 1)
print(classified_matrix)
}

}
