% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer_as_matrix.R
\name{get_layer_as_matrix}
\alias{get_layer_as_matrix}
\title{Get Layer as Matrix}
\usage{
get_layer_as_matrix(result_array, layer_index)
}
\arguments{
\item{result_array}{Un tableau 3D contenant plusieurs couches de données.}

\item{layer_index}{Un entier indiquant l'index de la couche à extraire. Doit être compris entre 1 et le nombre de couches dans `result_array`.}
}
\value{
Une matrice 2D représentant la couche spécifiée du tableau 3D.
}
\description{
Cette fonction extrait une couche spécifique d'un tableau 3D et la retourne sous forme de matrice 2D.
}
\examples{
# Création d'un tableau 3D pour l'exemple
result_array <- array(runif(300), dim = c(10, 10, 3))
# Extraction de la première couche sous forme de matrice
matrice <- get_layer_as_matrix(result_array, layer_index = 1)
print(matrice)
}
