% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3D_surface.R
\name{plot_3D_surface}
\alias{plot_3D_surface}
\title{Plot 3D Surface}
\usage{
plot_3D_surface(layer_index, plot_title)
}
\arguments{
\item{layer_index}{Un entier indiquant l'index de la couche à tracer. Doit être compris entre 1 et le nombre de couches dans `result_array`.}

\item{plot_title}{Un titre pour le graphique.}
}
\value{
Un graphique 3D interactif de la surface pour la couche spécifiée.
}
\description{
Cette fonction génère un graphique 3D interactif de la surface pour une couche spécifiée dans un tableau 3D.
}
\examples{
# Exemple d'utilisation de plot_3D_surface
result_array <- array(runif(100), dim = c(10, 10, 5))
plot_3D_surface(layer_index = 1, plot_title = "3D Surface Plot")
}
