% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Calculate and Plot Average Mg/Si Ratio by Region}
\usage{
region(index1, index2)
}
\arguments{
\item{index1}{Integer. The index of the layer in `result_array` that defines the regions.}

\item{index2}{Integer. The index of the layer in `result_array` containing Mg/Si values.}
}
\value{
A Plotly object. The function returns an interactive bar chart displaying the average Mg/Si ratio by region.
}
\description{
This function computes the average magnesium to silicon (Mg/Si) ratio for different regions defined in a matrix, and plots the results using Plotly.
}
\details{
The function extracts the specified region and Mg/Si layers from `result_array` using the provided indices. It calculates the average Mg/Si ratio for each region defined in the `regions_matrix` and plots these averages in a bar chart using Plotly. The regions are labeled from 0 to 6, and the plot provides a visual representation of the average Mg/Si ratio for each region.
}
\examples{
\dontrun{
# Assuming `result_array` is a 3D array with suitable data
plot <- region(result_array, index1 = 1, index2 = 2)
print(plot)
}

}
